moved the slider to external component
